requirements.txt
annotated-types==0.6.0
anyio==4.3.0
av==11.0.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
ctranslate2==4.2.1
distro==1.9.0
faster-whisper==1.0.1
filelock==3.14.0
flatbuffers==24.3.25
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
mpmath==1.3.0
numpy==1.26.4
onnxruntime==1.17.3
openai==1.22.0
packaging==24.0
protobuf==5.26.1
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
Pygments==2.17.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
setuptools==69.5.1
sniffio==1.3.1
sympy==1.12
tokenizers==0.15.2
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
